Base support for region and progress fields
Add fields for marking where an event came from and how far along it is,
for future use in filtering and/or display.

+++ feeds/common.go
@@ -26,10 +26,10 @@ func eventFilter(
 	event *perspective.EventData,
 	minTime int32,
 	maxTime int32,
-	typeFilter int16,
+	typeFilter int,
 	statusFilter int) bool {
 	if minTime < event.Start && maxTime > event.Start {
-		if typeFilter < 0 || event.Type == typeFilter {
+		if typeFilter < 0 || int(event.Type) == typeFilter {
 			if event.Status == 0 && 4&statusFilter != 0 {
 				return true // Done
 			}
